Resolve Unit Test Issues
This test is currently ignored because the SL61 Jenkins build machine is
failing when creating the Geotools-EPSG-HSQL database during Unit Test
time. The GeoTools library cannot create the proper database, thus the
library fails to find lookup codes, thus this unit test fails because
the data can't be parsed properly. When the SL61 issue is resolved, this
Unit test MUST be re-included into the suite.